--tag-one option to tag one host from a pool
Sometimes you want to use Rundeck to run a command on any one of a pool
of identical machines, rather than across all of them. To support this,
the --tag-one option applies an additional tag to one of the nodes
that supports a given service, pseudo-randomly.

In rundeck then one can specify a node filter containing that tag and
get a single randomly-selected host to run a job on.